otakusan: convert to a template
