+++ src/rust/otakusan/Cargo.toml
@@ -1,0 +1,11 @@
+[workspace]
+members = ["template", "sources/*"]
+
+[profile.dev]
+panic = "abort"
+
+[profile.release]
+panic = "abort"
+opt-level = "s"
+strip = true
+lto = true
