+++ src/rust/otakusan/res/filters.json
