+++ src/rust/otakusan/sources/myrockmanga/Cargo.toml
@@ -1,19 +1,11 @@
 [package]
-name = "otaku_sanctuary"
+name = "myrockmanga"
 version = "0.1.0"
 edition = "2021"
 
 [lib]
 crate-type = ["cdylib"]
 
-[profile.dev]
-panic = "abort"
-
-[profile.release]
-panic = "abort"
-opt-level = "s"
-strip = true
-lto = true
-
 [dependencies]
 aidoku = { git = "https://github.com/Aidoku/aidoku-rs" }
+otaku_sanctuary_template = { path = "../../template" }
