+++ src/rust/otakusan/sources/myrockmanga/res/source.json
@@ -1,0 +1,57 @@
+{
+	"info": {
+		"id": "multi.myrockmanga",
+		"lang": "multi",
+		"name": "MyRockManga",
+		"version": 1,
+		"url": "https://myrockmanga.com",
+		"nsfw": 2
+	},
+	"listings": [
+		{
+			"name": "Cosplay"
+		},
+		{
+			"name": "Wallpaper"
+		},
+		{
+			"name": "Ecchi Land"
+		},
+		{
+			"name": "For Boys"
+		},
+		{
+			"name": "For Girls"
+		},
+		{
+			"name": "New Titles"
+		},
+		{
+			"name": "Completed"
+		}
+	],
+	"languages": [
+		{
+			"code": "vi", 
+			"value": "vn",
+			"default": true
+		},
+		{
+			"code": "en",
+			"value": "us"
+		},
+		{
+			"code": "fr"
+		},
+		{
+			"code": "es"
+		},
+		{
+			"code": "it"
+		},
+		{
+			"code": "All"
+		}
+	],
+	"languageSelectType": "single"
+}
