+++ src/rust/otakusan/sources/otakusan/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "otakusan"
+version = "0.1.0"
+edition = "2021"
+
+[lib]
+crate-type = ["cdylib"]
+
+[dependencies]
+aidoku = { git = "https://github.com/Aidoku/aidoku-rs" }
+otaku_sanctuary_template = { path = "../../template" }
