+++ src/rust/otakusan/sources/otakusan/res/Icon.png
