+++ src/rust/otakusan/sources/otakusan/res/source.json
