+++ src/rust/otakusan/template/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "otaku_sanctuary_template"
+version = "0.1.0"
+edition = "2021"
+publish = false
+
+[dependencies]
+aidoku = { git = "https://github.com/Aidoku/aidoku-rs/" }
