+++ src/rust/otakusan/template/src/lib.rs
@@ -1,0 +1,6 @@
+#![no_std]
+extern crate alloc;
+pub mod helper;
+pub mod template;
+
+
